requirements.txt
torch
torchvision
pyyaml
accelerate
transformers
diffusers
peft